rewrote job_counter to not use pandas
